(web admin) replace component
